refactorization: Reduced __main__.py with 500 lines - created separate project module
